break logging functionality to another module
